Changed LastModificationTime and Size to 64-bit integers in LuaFile.

+++ lua/carrier/packages/panopticon.verification/luafile.lua
@@ -39,8 +39,8 @@ end
 function self:Serialize (streamWriter)
 	streamWriter:StringN16 (self.Path)
 	streamWriter:UInt32 (self.CRC32)
-	streamWriter:UInt32 (self.LastModificationTime)
-	streamWriter:UInt32 (self.Size)
+	streamWriter:UInt64 (self.LastModificationTime)
+	streamWriter:UInt64 (self.Size)
 	streamWriter:Boolean (self.DynamicCode)
 	streamWriter:UInt32 (self.FunctionCount)
 	
@@ -59,8 +59,8 @@ end
 function self:Deserialize (streamReader)
 	self.Path                 = streamReader:StringN16 ()
 	self.CRC32                = streamReader:UInt32 ()
-	self.LastModificationTime = streamReader:UInt32 ()
-	self.Size                 = streamReader:UInt32 ()
+	self.LastModificationTime = streamReader:UInt64 ()
+	self.Size                 = streamReader:UInt64 ()
 	self.DynamicCode          = streamReader:Boolean ()
 	self.FunctionCount        = streamReader:UInt32 ()
